wip: separate concerns of spec execution specifics into runtime, colocate pipe runtime ctx
